grpc transport

+++ internal/util/math.go
@@ -1,6 +1,6 @@
 package util
 
-func Min(x, y int) int {
+func Min(x, y uint64) uint64 {
 	if x < y {
 		return x
 	}
